pyproject.toml
[project]
name = "image-to-ico"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "numpy>=2.0.2",
    "pillow>=11.3.0",
]
